fix 64bit issues

+++ common/GDGradient.h
@@ -5,17 +5,17 @@
 #import <Cocoa/Cocoa.h>
 
 
-@interface GDGradient : NSObject
-{
+@interface GDGradient : NSObject {
     CGColorSpaceRef m_colorSpace;
     CGFunctionRef   m_shadingFunction;
     NSColor*        m_color;
-    float           m_highlight[4];
-    float           m_shadow[4];
+    CGFloat         m_highlight[4];
+    CGFloat         m_shadow[4];
 }
+
 - (NSColor *)color;
 - (void)setColor:(NSColor *)color;
-- (void)getGradientColorComponents:(float[4])components forFraction:(float)fraction;
-- (NSColor *)gradientColorForFraction:(float)fraction;
+- (void)getGradientColorComponents:(CGFloat[4])components forFraction:(CGFloat)fraction;
+- (NSColor *)gradientColorForFraction:(CGFloat)fraction;
 - (void)fillRect:(NSRect)aRect;
 @end
